Remove duplicated code
